Aligned Popover Story, removed navscroll
