--- stories/NavScroll/NavScroll.stories.mdx
@@ -1,41 +1,0 @@
-import {
-  ArgsTable,
-  Meta,
-  Canvas,
-  Story,
-  Preview
-} from '@storybook/addon-docs/blocks';
-import {
-  // Components to include
-  Header,
-  Headers,
-  HeaderContent,
-  HeaderLinkZone,
-  HeaderRightZone,
-  HeaderBrand,
-  HeaderSearch,
-  HeaderSocialsZone,
-  HeaderToggler
-} from '../../src';
-
-<Meta title='Documentazione/Menu di navigazione/NavScroll' of={Header} />
-
-# NavScroll
-
-## Posizionamento a fondo pagina con linea a sinistra
-
-<Canvas>
-  <Story id='componenti-navscroll--posizionamento-a-fondo-pagina-con-linea-a-sinistra' />
-</Canvas>
-
-## Posizionamento in testa alla pagina con linea a destra
-
-<Canvas>
-  <Story id='componenti-navscroll--posizionamento-in-testa-alla-pagina-con-linea-a-destra' />
-</Canvas>
-
-## Menu Inline
-
-<Canvas>
-  <Story id='componenti-navscroll--menu-inline' />
-</Canvas>
